+++ src/main/java/com/kakao/saramaracommunity/comment/dto/business/request/CommentDeleteServiceRequest.java
@@ -1,0 +1,9 @@
+package com.kakao.saramaracommunity.comment.dto.business.request;
+
+import java.util.Objects;
+import lombok.Builder;
+
+@Builder
+public record CommentDeleteServiceRequest(Long memberId) {
+
+}
