+++ src/main/java/com/kakao/saramaracommunity/comment/dto/business/response/CommentsReadInBoardResponse.java
@@ -1,4 +1,4 @@
-package com.kakao.saramaracommunity.comment.service.dto.response;
+package com.kakao.saramaracommunity.comment.dto.business.response;
 
 import com.kakao.saramaracommunity.comment.entity.Comment;
 import java.util.List;
