+++ src/main/java/com/kakao/saramaracommunity/comment/service/CommentService.java
@@ -1,10 +1,10 @@
 package com.kakao.saramaracommunity.comment.service;
 
-import com.kakao.saramaracommunity.comment.service.dto.request.CommentCreateServiceRequest;
-import com.kakao.saramaracommunity.comment.service.dto.request.CommentDeleteServiceRequest;
-import com.kakao.saramaracommunity.comment.service.dto.request.CommentUpdateServiceRequest;
-import com.kakao.saramaracommunity.comment.service.dto.response.CommentCreateResponse;
-import com.kakao.saramaracommunity.comment.service.dto.response.CommentsReadInBoardResponse;
+import com.kakao.saramaracommunity.comment.dto.business.request.CommentCreateServiceRequest;
+import com.kakao.saramaracommunity.comment.dto.business.request.CommentDeleteServiceRequest;
+import com.kakao.saramaracommunity.comment.dto.business.request.CommentUpdateServiceRequest;
+import com.kakao.saramaracommunity.comment.dto.business.response.CommentCreateResponse;
+import com.kakao.saramaracommunity.comment.dto.business.response.CommentsReadInBoardResponse;
 
 public interface CommentService {
