+++ src/main/java/com/kakao/saramaracommunity/comment/service/CommentServiceImpl.java
@@ -6,11 +6,11 @@ import com.kakao.saramaracommunity.comment.entity.Comment;
 import com.kakao.saramaracommunity.comment.exception.CommentBusinessException;
 import com.kakao.saramaracommunity.comment.exception.CommentErrorCode;
 import com.kakao.saramaracommunity.comment.repository.CommentRepository;
-import com.kakao.saramaracommunity.comment.service.dto.request.CommentCreateServiceRequest;
-import com.kakao.saramaracommunity.comment.service.dto.request.CommentDeleteServiceRequest;
-import com.kakao.saramaracommunity.comment.service.dto.request.CommentUpdateServiceRequest;
-import com.kakao.saramaracommunity.comment.service.dto.response.CommentCreateResponse;
-import com.kakao.saramaracommunity.comment.service.dto.response.CommentsReadInBoardResponse;
+import com.kakao.saramaracommunity.comment.dto.business.request.CommentCreateServiceRequest;
+import com.kakao.saramaracommunity.comment.dto.business.request.CommentDeleteServiceRequest;
+import com.kakao.saramaracommunity.comment.dto.business.request.CommentUpdateServiceRequest;
+import com.kakao.saramaracommunity.comment.dto.business.response.CommentCreateResponse;
+import com.kakao.saramaracommunity.comment.dto.business.response.CommentsReadInBoardResponse;
 import com.kakao.saramaracommunity.member.entity.Member;
 import com.kakao.saramaracommunity.member.repository.MemberRepository;
 import java.util.List;
