+++ src/test/java/com/kakao/saramaracommunity/comment/service/CommentServiceImplTest.java
@@ -12,14 +12,14 @@ import com.kakao.saramaracommunity.board.entity.Board;
 import com.kakao.saramaracommunity.board.entity.CategoryBoard;
 import com.kakao.saramaracommunity.board.repository.BoardImageRepository;
 import com.kakao.saramaracommunity.board.repository.BoardRepository;
-import com.kakao.saramaracommunity.comment.controller.dto.request.CommentCreateRequest;
-import com.kakao.saramaracommunity.comment.controller.dto.request.CommentDeleteRequest;
-import com.kakao.saramaracommunity.comment.controller.dto.request.CommentUpdateRequest;
+import com.kakao.saramaracommunity.comment.dto.api.request.CommentCreateRequest;
+import com.kakao.saramaracommunity.comment.dto.api.request.CommentDeleteRequest;
+import com.kakao.saramaracommunity.comment.dto.api.request.CommentUpdateRequest;
 import com.kakao.saramaracommunity.comment.entity.Comment;
 import com.kakao.saramaracommunity.comment.exception.CommentBusinessException;
 import com.kakao.saramaracommunity.comment.repository.CommentRepository;
-import com.kakao.saramaracommunity.comment.service.dto.response.CommentCreateResponse;
-import com.kakao.saramaracommunity.comment.service.dto.response.CommentsReadInBoardResponse;
+import com.kakao.saramaracommunity.comment.dto.business.response.CommentCreateResponse;
+import com.kakao.saramaracommunity.comment.dto.business.response.CommentsReadInBoardResponse;
 import com.kakao.saramaracommunity.member.entity.Member;
 import com.kakao.saramaracommunity.member.repository.MemberRepository;
 import com.kakao.saramaracommunity.support.IntegrationTestSupport;
